refactor: 오류 메세지 핸들러 수정
refactor: 오류 메세지 핸들러 수정
